+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+### 1.2.0 (2021-06-22)
+* Automatic website logo checker added.
+* Primary website image color as image background.
+* Random color getting when url is invalid or image not found.
+
+
 ### 1.1.0 (2021-06-21)
 * Password generator feature added.